feat(qjs_v8_compat): PromiseResolver resolve/reject accept Into<Local<Value>>
Allows callers to pass Local<Object> etc. directly without explicit
.into(). Also adds CachedData Deref<Target=[u8]>/AsRef so deno_core's
ready_callback that takes &[u8] works on &Box<CachedData>.
FixedArray::length returns usize. compile_module takes &mut Source.

Drops deno_core --features quickjs build errors from 264 to 261.

+++ libs/qjs_v8_compat/src/module.rs
@@ -153,7 +153,7 @@ impl<'s> Local<'s, ModuleRequest> {
 }
 
 impl<'s> Local<'s, FixedArray> {
-  pub fn length(&self) -> i32 {
+  pub fn length(&self) -> usize {
     0
   }
   pub fn get<S>(
